backend/requirements.txt
tenacity
pydantic-settings
fastapi
uvicorn[standard]
python-dotenv==1.0.1
reportlab==4.2.5
PyPDF2==3.0.1
pydantic
python-multipart==0.0.20
openai==1.57.2
duckduckgo-search>=8.1.1
yfinance==0.2.66
supabase
feedparser
pandas
langfuse>=2.0.0
pdfplumber>=0.10.0
pypdfium2>=4.0.0
Pillow>=10.0.0
